pyproject.toml
[tool.poetry]
name = "kinit-fast-task"
version = "5.0.0"
description = "FastAPI 技术栈最佳实践"
authors = ["ktianc <2445667550@qq.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://python-poetry.org/"
repository = "https://github.com/python-poetry/poetry"
documentation = "https://python-poetry.org/docs/main/pyproject/"
packages = [
    { include = "kinit_fast_task" },
]


[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.2.1"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.29"
pydantic = "^2.6.4"
loguru = "^0.7.2"
user-agents = "^2.2.0"
redis = "^5.0.3"
pycryptodome = "^3.20.0"
alibabacloud-tea-openapi = "^0.3.8"
alibabacloud-dysmsapi20170525 = "^2.0.24"
oss2 = "^2.18.4"
aiopathlib = "^0.5.0"
aioshutil = "^1.3"
openpyxl = "^3.1.2"
xlsxwriter = "^3.2.0"
uvicorn = "^0.29.0"
typer = "^0.12.3"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
motor = "^3.4.0"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-html = "^4.1.1"
pytest-dotenv = "^0.5.2"
ruff = "^0.4.1"
pre-commit = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"